requirements.txt
# Core
requests>=2.31.0
feedparser>=6.0.11
pandas>=2.1.0
SQLAlchemy>=2.0.0

# Scheduling
schedule>=1.2.0

# Streamlit & Visualization
streamlit>=1.32.0
plotly>=5.19.0

# Data Validation
pydantic>=2.5.0

# Logging
loguru>=0.7.2

# Environment
python-dotenv>=1.0.0

# Optional ML
scikit-learn>=1.4.0

# NLP / Text
thefuzz>=0.22.1
python-Levenshtein>=0.23.0
